Optimize cython further by eliminating bounds checking

+++ searchlight.pyx
@@ -12,11 +12,11 @@ def searchlight(DTYPE_t[:, ::1] cc1,
                 DTYPE_t[:, ::1] cc2,
                 DTYPE_t[:] f):
     assert cc1.shape[1] == cc2.shape[1] == f.shape[0]
-    cdef int sum = 0
-    cdef Py_ssize_t ndims = cc1.shape[1]
-    cdef Py_ssize_t ncc1 = cc1.shape[0]
-    cdef Py_ssize_t ncc2 = cc2.shape[0]
-    cdef Py_ssize_t i1, i2, dim
+    cdef unsigned long sum = 0
+    cdef size_t ndims = cc1.shape[1]
+    cdef size_t ncc1 = cc1.shape[0]
+    cdef size_t ncc2 = cc2.shape[0]
+    cdef size_t i1, i2, dim
     with nogil:
         for i1 in prange(ncc1, schedule='guided'):
             for i2 in range(ncc2):
